// src/types/textEdit.ts
export type TextEditFormatCommand =
  | 'bold'
  | 'italic'
  | 'strikethrough'
  | 'underline'
  | 'superscript'
  | 'subscript';

export interface TextEditFormatPayload {
  command: TextEditFormatCommand;
}

export interface TextEditFontPayload {
  fontFamily: string;
}

export interface TextEditFontSizePayload {
  fontSize: number;
}

export interface TextEditFontOptionsResponse {
  fonts: string[];
  currentFont?: string | null;
  currentFontSize?: number | null;
}